+++ src/main/java/com/minsoo/co/tireerpserver/model/response/tire/TireResponse.java
@@ -26,7 +26,7 @@ public class TireResponse {
     private String oe;
     private String originalVehicle;
     private String countryOfManufacture;
-    private Integer retailPrice;
+    private Long retailPrice;
 
     private String tireGroup;
     private String note;
